Refactor Character Creation and move appropriate functions to TestCenter.cs
